refactor(interpreter): split env.rs into env / value / artifact modules
Move the Value enum to src/value.rs and the artifact types
(ArtifactMethod, ArtifactSchema, ArtifactFieldSchema, ArtifactValue,
ArtifactHandle) to src/artifact.rs, leaving env.rs with RuntimeEnv,
the callable machinery, and VarInfo. env.rs re-exports the moved
types so all pre-split paths keep working. The env unit tests reuse
eval::test_support::rune instead of a local copy.

Closes #504
