+++ crates/abyss-interpreter/src/lib.rs
@@ -1,5 +1,7 @@
+pub mod artifact;
 pub(crate) mod diagnostics;
 pub mod env;
 pub mod eval;
 pub mod io_bridge;
 pub mod stdlib;
+pub mod value;
